Create global error handler
