+++ api-webflux/src/main/kotlin/com/rafael/apiwebflux/Application.kt
@@ -1,6 +1,5 @@
 package com.rafael.apiwebflux
 
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.runApplication
