+++ api-webflux/src/test/kotlin/com/rafael/apiwebflux/http/EliigibleControllerTest.kt
@@ -12,7 +12,6 @@ import com.rafael.models.Eligible
 import com.rafael.models.SearchResult
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import org.mockito.Mockito.`when`
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest
 import org.springframework.boot.test.mock.mockito.MockBean
@@ -21,6 +20,7 @@ import org.springframework.test.web.reactive.server.WebTestClient
 import org.springframework.test.web.reactive.server.expectBody
 import retrofit2.Call
 import retrofit2.Response
+import org.mockito.Mockito.`when`
 
 @ContextConfiguration(
     classes = [
